Setup python connector to MySQL database
The connector allows interaction with the MySQL database through
python. The current functionality only allows for retrieving all data
from a specified table, or inserting a single row into a table. The
connector relies on a config file which specifies the credentials used
to access the MySQL database. The default constructor expects the file
to be called 'config.ini' but a different file can be specified. The
config file was not included in this commit for security reasons, but
it should have the following form:

[mysql]
host = my_host
database = my_database
user = my_user
password = password_for_my_user